// src/constants/interfaces.ts
export interface Interface__Route {
  path: string;
  activePath: string;
  element: any;
}
export interface Interface__PrivateRoute extends Interface__Route {
  label?: any;
  titleKey: string;
  backPath?: string;
  permissions?: number[];
  allowedPermissions?: number[];
}